Реализация больших чисел

+++ app/src/main/java/com/infinity_coder/divcalendar/data/db/model/PaymentDbModel.kt
@@ -9,6 +9,7 @@ import com.infinity_coder.divcalendar.data.db.model.PaymentDbModel.Companion.COL
 import com.infinity_coder.divcalendar.data.db.model.PaymentDbModel.Companion.INDEX_SECURITY
 import com.infinity_coder.divcalendar.data.db.model.PaymentDbModel.Companion.TABLE_NAME
 import com.infinity_coder.divcalendar.data.network.model.PaymentNetModel
+import java.math.BigDecimal
 
 @Entity(
     tableName = TABLE_NAME,
@@ -23,7 +24,7 @@ import com.infinity_coder.divcalendar.data.network.model.PaymentNetModel
 )
 data class PaymentDbModel(
     @ColumnInfo(name = COLUMN_DIVIDENDS)
-    var dividends: Double,
+    var dividends: BigDecimal,
 
     @ColumnInfo(name = COLUMN_DATE)
     val date: String,
@@ -41,7 +42,7 @@ data class PaymentDbModel(
     var exchange: String = "",
 
     @ColumnInfo(name = COLUMN_COUNT)
-    var count: Int? = null
+    var count: BigDecimal? = null
 ) {
     @Ignore
     var security: SecurityDbModel? = null
@@ -60,7 +61,7 @@ data class PaymentDbModel(
         const val INDEX_SECURITY = "security_index"
 
         fun from(portfolioId: Long, exchange: String, networkPayments: PaymentNetModel.Response) = PaymentDbModel(
-            dividends = networkPayments.dividends,
+            dividends = BigDecimal(networkPayments.dividends),
             date = networkPayments.date,
             forecast = networkPayments.forecast,
             isin = if (networkPayments.isin.isBlank()) networkPayments.name else networkPayments.isin,
